MCP 服务移入 services/ 目录，所有工具调用增加日志与异常追踪
